Fix repositories order
Due to recent changes, Android is no longer supporting downloading support libraries through the SDK Manager. The support libraries are now available through Google's Maven repository.

This leads to not be able to install some plugins which are up to date with this rule.

For some reasons, just inverting the repository search order will fix the problem. As seen in https://github.com/meteor/meteor/issues/10308

This PR does the following:

- inverse the repositories order